test(core): split hmr incremental graph tests
