+++ packages/core/test/core/hmr-perf/incremental-graph/stale-content-hash.ts
@@ -1,0 +1,50 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../helpers';
+
+describe('Issue #1223 HMR perf - stale content hash cleanup', () => {
+  test('phase1d: import 제거 후 이전 파일 변경은 리빌드 트리거 안 함', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-1223-phase1d-'));
+    const entry = join(dir, 'entry.ts');
+    const extra = join(dir, 'extra.ts');
+    writeFileSync(extra, 'export const y = 1;');
+    writeFileSync(entry, 'import { y } from "./extra"; export const x = y;');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const rebuilds: Array<{ changed?: string[] }> = [];
+    const handle = watch({
+      entryPoints: [entry],
+      onReady() {
+        readyDone();
+      },
+      onRebuild(event) {
+        rebuilds.push(event);
+      },
+    });
+    await readyP;
+    await new Promise((r) => setTimeout(r, 100));
+
+    // 1차: entry에서 extra import 제거 → graph에서 extra 빠짐
+    writeFileSync(entry, 'export const x = 1;');
+    await new Promise((r) => setTimeout(r, 1500));
+    const reb1 = rebuilds.length;
+    expect(reb1).toBeGreaterThanOrEqual(1);
+
+    // 2차: extra.ts 내용 변경 — 이미 그래프에서 빠졌으므로 리빌드 없어야 함
+    writeFileSync(extra, 'export const y = 999;');
+    await new Promise((r) => setTimeout(r, 1500));
+    handle.stop();
+    rmSync(dir, { recursive: true });
+
+    // extra 변경 후 추가 리빌드가 없어야 — watcher가 extra를 removePath 한 결과
+    expect(rebuilds.length).toBe(reb1);
+  }, 15000);
+});
